test: move DMA specific tests to test_dma.py
